+++ 00_capi_bench/src/algorithms/capi/SSSP.c
@@ -266,7 +266,7 @@ uint32_t SSSPCompareDistanceArrays(struct SSSPStats *stats1, struct SSSPStats *s
 
 }
 
-int SSSPAtomicRelax(uint32_t src, uint32_t dest, uint32_t weight, struct SSSPStats *stats)
+int SSSPAtomicRelax(uint32_t src, uint32_t dest, float weight, struct SSSPStats *stats)
 {
     uint32_t oldParent, newParent;
     uint32_t oldDistanceV = UINT_MAX / 2;
@@ -344,7 +344,7 @@ int SSSPAtomicRelax(uint32_t src, uint32_t dest, uint32_t weight, struct SSSPSta
 
 
 
-int SSSPRelax(uint32_t src, uint32_t dest, uint32_t weight, struct SSSPStats *stats)
+int SSSPRelax(uint32_t src, uint32_t dest, float weight, struct SSSPStats *stats)
 {
 
 
@@ -450,7 +450,7 @@ void SSSPSpiltGraphCSR(struct GraphCSR *graph, struct GraphCSR **graphPlus, stru
     // uint32_t numVerticesPlusCounter = 0;
     // uint32_t numVerticesMinusCounter = 0;
     uint32_t e;
-    uint32_t weight;
+    float weight;
     // uint32_t src;
     // uint32_t dest;
 
@@ -1021,7 +1021,7 @@ struct SSSPStats *SSSPDataDrivenPushGraphCSR(uint32_t source,  uint32_t iteratio
                     {
                         uint32_t src = graphLight->sorted_edges_array->edges_array_src[j];
                         uint32_t dest = graphLight->sorted_edges_array->edges_array_dest[j];
-                        uint32_t weight = graphLight->sorted_edges_array->edges_array_weight[j];
+                        float weight = graphLight->sorted_edges_array->edges_array_weight[j];
 
                         if(numThreads == 1)
                             activeVertices += SSSPRelax(src, dest, weight, stats);
@@ -1055,7 +1055,7 @@ struct SSSPStats *SSSPDataDrivenPushGraphCSR(uint32_t source,  uint32_t iteratio
                 {
                     uint32_t src = graphHeavy->sorted_edges_array->edges_array_src[j];
                     uint32_t dest = graphHeavy->sorted_edges_array->edges_array_dest[j];
-                    uint32_t weight = graphHeavy->sorted_edges_array->edges_array_weight[j];
+                    float weight = graphHeavy->sorted_edges_array->edges_array_weight[j];
 
                     if(numThreads == 1)
                         activeVertices += SSSPRelax(src, dest, weight, stats);
