+++ 00_capi_bench/src/algorithms/capi/bellmanFord.c
@@ -200,7 +200,7 @@ uint32_t bellmanFordCompareDistanceArrays(struct BellmanFordStats *stats1, struc
 
 }
 
-int bellmanFordAtomicRelax(uint32_t src, uint32_t dest, uint32_t weight, struct BellmanFordStats *stats, struct Bitmap *bitmapNext)
+int bellmanFordAtomicRelax(uint32_t src, uint32_t dest, float weight, struct BellmanFordStats *stats, struct Bitmap *bitmapNext)
 {
     // uint32_t oldParent, newParent;
     uint32_t oldDistanceV = UINT_MAX / 2;
@@ -270,7 +270,7 @@ int bellmanFordAtomicRelax(uint32_t src, uint32_t dest, uint32_t weight, struct 
 
 
 
-int bellmanFordRelax(uint32_t src, uint32_t dest, uint32_t weight, struct BellmanFordStats *stats, struct Bitmap *bitmapNext)
+int bellmanFordRelax(uint32_t src, uint32_t dest, float weight, struct BellmanFordStats *stats, struct Bitmap *bitmapNext)
 {
 
     uint32_t activeVertices = 0;
@@ -484,7 +484,7 @@ struct BellmanFordStats *bellmanFordPullRowGraphGrid(uint32_t source,  uint32_t 
 
                     uint32_t src = partition->edgeList->edges_array_src[k];
                     uint32_t dest = partition->edgeList->edges_array_dest[k];
-                    uint32_t weight = partition->edgeList->edges_array_weight[k];
+                    float weight = partition->edgeList->edges_array_weight[k];
 
                     if(getBit(bitmapCurr, src))
                     {
@@ -604,7 +604,7 @@ struct BellmanFordStats *bellmanFordPushColumnGraphGrid(uint32_t source,  uint32
 
                     uint32_t src = partition->edgeList->edges_array_src[k];
                     uint32_t dest = partition->edgeList->edges_array_dest[k];
-                    uint32_t weight = partition->edgeList->edges_array_weight[k];
+                    float weight = partition->edgeList->edges_array_weight[k];
 
                     if(getBit(bitmapCurr, src))
                     {
@@ -1023,7 +1023,7 @@ struct BellmanFordStats *bellmanFordDataDrivenPushGraphCSR(uint32_t source,  uin
                 {
                     uint32_t src = graph->sorted_edges_array->edges_array_src[j];
                     uint32_t dest = graph->sorted_edges_array->edges_array_dest[j];
-                    uint32_t weight = graph->sorted_edges_array->edges_array_weight[j];
+                    float weight = graph->sorted_edges_array->edges_array_weight[j];
 
                     if(numThreads == 1)
                         activeVertices += bellmanFordRelax(src, dest, weight, stats, bitmapNext);
@@ -1193,7 +1193,7 @@ struct BellmanFordStats *bellmanFordRandomizedDataDrivenPushGraphCSR(uint32_t so
 
                     uint32_t src = graphPlus->sorted_edges_array->edges_array_src[j];
                     uint32_t dest = graphPlus->sorted_edges_array->edges_array_dest[j];
-                    uint32_t weight = graphPlus->sorted_edges_array->edges_array_weight[j];
+                    float weight = graphPlus->sorted_edges_array->edges_array_weight[j];
 
                     if(numThreads == 1)
                         activeVertices += bellmanFordRelax(src, dest, weight, stats, bitmapNext);
@@ -1220,7 +1220,7 @@ struct BellmanFordStats *bellmanFordRandomizedDataDrivenPushGraphCSR(uint32_t so
 
                     uint32_t src = graphMinus->sorted_edges_array->edges_array_src[j];
                     uint32_t dest = graphMinus->sorted_edges_array->edges_array_dest[j];
-                    uint32_t weight = graphMinus->sorted_edges_array->edges_array_weight[j];
+                    float weight = graphMinus->sorted_edges_array->edges_array_weight[j];
 
                     if(numThreads == 1)
                         activeVertices += bellmanFordRelax(src, dest, weight, stats, bitmapNext);
@@ -1534,7 +1534,7 @@ struct BellmanFordStats *bellmanFordDataDrivenPushGraphAdjArrayList(uint32_t sou
 
                     uint32_t src = nodes->edges_array_src[j];
                     uint32_t dest = nodes->edges_array_dest[j];
-                    uint32_t weight = nodes->edges_array_weight[j];
+                    float weight = nodes->edges_array_weight[j];
 
                     if(numThreads == 1)
                         activeVertices += bellmanFordRelax(src, dest, weight, stats, bitmapNext);
